Rearrange GPB .proto files
To closer reflect the services provided by etcd.

We lump health checks, cluster membership, alarms into
"maintenance" as they are arguably of less interest to the
users of this client (compared to the K/V, leases, auth
and locking services).

This also introduces a lock.proto to lay a foundation
for the future work on #14.