Refactor eetcd_conn module
Since the eetcd provides named clients, it is not necessary to maintain the named processes manually. Use `one_for_one` strategy supervisor instead of `simple_one_for_one`. Let the local register manage the named clients.

Since the `restart` is set to `transient`, the `eetcd_conn` process should always be restarted when it crashes or exits abnormally.

Use in-process state round-robin selection instead of ETS table. For etcd clients, API calls should not be made too frequently.

The previous `eetcd_conn` module uses `gen_statem` behaviour, and the error handling will block and cannot be stopped from the outside. For example, when etcd nodes restart, this process will make thousands of retries, which may reach the intensity of the restart settings. That is because the gen_statem next event loop is inserted in front of its mailbox.

Further thought, this etcd client was built on top of gun, an HTTP client, directly instead of a full-featured gRPC framework. That makes the code too complicated to handle the gRPC semantics and behaviours. For example, there should be interceptors to do retry API calls, authentication, token refresh, or balancer tasks.